+++ README.md
@@ -4,7 +4,7 @@ This repository contains the necessary tools to setup, train, and test a neural 
 
 ## Setup
 
-Pull down the model onto the server. Open up `model.py` and make any necessary changes to the global parameters. You can add hidden layers if you want, or just run it as-is. You can also change the data subset used to train / validate.
+Two models are currently available for testing: ResNet50 and EfficientNetB7. They are run with their respective files using Python3. At the top of the file, you may specify parameters such as hidden layers, and whether or not you want a plot of the training performance over time.
 
 ## Training the Model
 
@@ -12,7 +12,5 @@ Ensure you are using the correct Python3 environment. The server uses `conda` to
 ```
 conda activate tensorflow
 ```
-This environment should have all the tools needed to run the model. After that, it's as simple as
-```
-python3 model.py
-```
+
+If you're using the EfficientNetB7 model, it is not contained in the conda environment. EfficientNet is only support on TensorFlow nightly builds, so this has been installed with Pip. If running on the server, it should already by accessible to you.
